ai_service/requirements.txt
fastapi
uvicorn
python-multipart
torch==2.6.0+cpu
torchvision==0.21.0
--extra-index-url https://download.pytorch.org/whl/cpu
pillow
numpy
ultralytics
opencv-python-headless
transformers
sentencepiece
requests
